db: store log file hash for later search
A problem with the current implementation of database storage of ulogs
is that different systems sharing a log database would need to
independently keep track of the primary_key of a given log. As an
example, one might want to use flight_review for log analysis, but also
get flight log statics directly into a fleet management system. This
could be solved by synchronization scripts, but this tightly couples the
systems with room for error.

This commit solves the problem by storing the SHA256 hash of the file in
the database, so that anyone posessing the binary file can look it up in
the database without having to know the primary key a priori.

This also solves the problem of duplicate uploads, since the hash is
constrained to be unique.