test: generalize test_ulog_write
Use __dict__ comparison of ULog objects instead of manually specifying
each field, since that is more future-proof. In general, we will want to
expect all fields of a written ulog to be equal to the original. This
unveiled a bug in the _incompat_flags part of write_ulog, see separate
commit on fix.